refactor: WarehouseVO 위치 이동
