+++ src/test/java/com/ssg/meowwms/mapper/warehouse/WarehouseMapperTest.java
@@ -1,15 +1,12 @@
 package com.ssg.meowwms.mapper.warehouse;
 
-import com.ssg.meowwms.domain.WarehouseVO;
+import com.ssg.meowwms.domain.warehouse.WarehouseVO;
 import com.ssg.meowwms.mapper.WarehouseMapper;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Transactional
